Fixed error with ValueNodes (Always having parents {-1, -1})

+++ CartesianGP/Extensions/DateTimeExtensions.cs
@@ -2,6 +2,6 @@ static class DateTimeExtensions
 {
     public static int GetTimestamp(this DateTime value)
     {
-        return int.Parse( value.ToString("yyyyMMddHHmmssffff") );
+        return (int)long.Parse( value.ToString("yyyyMMddHHmmssffff") );
     }
 }
